requirements.txt
Package                   Version
------------------------- --------------
absl-py                   2.3.1
anyio                     4.11.0
argon2-cffi               25.1.0
argon2-cffi-bindings      25.1.0
arrow                     1.3.0
asttokens                 3.0.0
astunparse                1.6.3
async-lru                 2.0.5
attrs                     25.3.0
babel                     2.17.0
beautifulsoup4            4.14.2
bleach                    6.2.0
blinker                   1.9.0
certifi                   2025.8.3
cffi                      2.0.0
charset-normalizer        3.4.3
click                     8.3.0
colorama                  0.4.6
coloredlogs               15.0.1
comm                      0.2.3
contourpy                 1.3.3
cycler                    0.12.1
debugpy                   1.8.17
decorator                 5.2.1
deepface                  0.0.95
defusedxml                0.7.1
entrypoints               0.4
executing                 2.2.1
fastjsonschema            2.21.2
filelock                  3.19.1
fire                      0.7.1
Flask                     3.1.2
flask-cors                6.0.1
flatbuffers               25.9.23
fonttools                 4.60.1
fqdn                      1.5.1
fsspec                    2025.9.0
gast                      0.6.0
gdown                     5.2.0
google-pasta              0.2.0
grpcio                    1.75.1
gunicorn                  23.0.0
h11                       0.16.0
h5py                      3.14.0
httpcore                  1.0.9
httpx                     0.28.1
humanfriendly             10.0
idna                      3.10
ipykernel                 6.29.5
ipython                   9.6.0
ipython_pygments_lexers   1.1.1
isoduration               20.11.0
itsdangerous              2.2.0
jedi                      0.19.2
Jinja2                    3.1.6
joblib                    1.5.2
json5                     0.12.1
jsonpointer               3.0.0
jsonschema                4.25.1
jsonschema-specifications 2025.9.1
jupyter_client            7.4.9
jupyter_core              5.8.1
jupyter-events            0.12.0
jupyter-lsp               2.3.0
jupyter_server            2.17.0
jupyter_server_terminals  0.5.3
jupyterlab                4.4.9
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
keras                     3.11.3
kiwisolver                1.4.9
lark                      1.3.0
libclang                  18.1.1
lz4                       4.4.4
Markdown                  3.9
markdown-it-py            4.0.0
MarkupSafe                3.0.3
matplotlib                3.10.6
matplotlib-inline         0.1.7
mdurl                     0.1.2
mistune                   3.1.4
ml_dtypes                 0.5.3
mpmath                    1.3.0
mtcnn                     1.0.0
namex                     0.1.0
nbclient                  0.10.2
nbconvert                 7.16.6
nbformat                  5.10.4
nest-asyncio              1.6.0
networkx                  3.3
notebook                  7.4.7
notebook_shim             0.2.4
numpy                     2.2.6
onnx                      1.19.0
onnxruntime               1.23.0
opencv-python             4.12.0.88
opt_einsum                3.4.0
optree                    0.17.0
overrides                 7.7.0
packaging                 25.0
pandas                    2.3.3
pandocfilters             1.5.1
parso                     0.8.5
pillow                    11.3.0
pip                       25.2
platformdirs              4.4.0
prometheus_client         0.23.1
prompt_toolkit            3.0.52
protobuf                  6.32.1
psutil                    7.1.0
pure_eval                 0.2.3
pycparser                 2.23
Pygments                  2.19.2
pyparsing                 3.2.5
pyreadline3               3.5.4
PySocks                   1.7.1
python-dateutil           2.9.0.post0
python-json-logger        3.3.0
pytz                      2025.2
pywin32                   311
pywinpty                  3.0.2
PyYAML                    6.0.3
pyzmq                     24.0.1
referencing               0.36.2
requests                  2.32.5
retina-face               0.0.17
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rfc3987-syntax            1.1.0
rich                      14.1.0
rpds-py                   0.27.1
scikit-learn              1.7.2
scipy                     1.16.2
Send2Trash                1.8.3
setuptools                65.5.0
six                       1.17.0
sniffio                   1.3.1
soupsieve                 2.8
stack-data                0.6.3
sympy                     1.13.3
tensorboard               2.20.0
tensorboard-data-server   0.7.2
tensorflow                2.20.0
termcolor                 3.1.0
terminado                 0.18.1
tf_keras                  2.20.1
threadpoolctl             3.6.0
tinycss2                  1.4.0
torch                     2.8.0+cu126
torchvision               0.23.0+cu126
tornado                   6.5.2
tqdm                      4.67.1
traitlets                 5.14.3
types-python-dateutil     2.9.0.20250822
typing_extensions         4.15.0
tzdata                    2025.2
uri-template              1.3.0
urllib3                   2.5.0
wcwidth                   0.2.14
webcolors                 24.11.1
webencodings              0.5.1
websocket-client          1.8.0
Werkzeug                  3.1.3
wheel                     0.45.1
wrapt                     1.17.3
